Remove ratings
Ratings hinder my real expressions about the book. I mostly tend to assume 4 for
a non-fiction and go to 5 if it /impressed/, or 3 if it bored me down. Plus I
pick up non-fictions which I find interesting beforehand, this mostly sets the
ratings high. This is a pretty shallow thing to leave for my future self.
Ratings are good indications for some things, but only when aggregated over a
vast set of readers.

I am going to keep this as a books + notes repo.